Update readme to include sorting the releases
Currently the releases are read in the way that git outputs them, which seems to be sorted alphabetical. If a project changed between the v0.0.0 and 0.0.0 format, this will result in the returned release not being sorted correctly.

The option is to just sort the after loading them, like I am including in the Readme here, or the other option would be to sort it in the versions method before returning.